Upgrade from upstream repo (#1)
* Replace assertTrue(isinstance()) with assertIsInstance()

Some of tests use different method of assertTrue(isinstance(A, B)) or
assertEqual(type(A), B). The correct way is to use assertIsInstance(A,
B) provided by testtools.

Change-Id: Ia92e870a27cc2fe4b7b07a7097410b2cb42184a0
Closes-Bug: #1268480

* Remove translation of log messages

The i18n team has decided not to translate the logs because it
seems like it not very useful; operators prefer to have them in
English so that they can search for those strings on the internet.

Change-Id: I84b1a2397f424430ed332aa7d4d874379abd0357

* Updated from global requirements

Change-Id: I679de7c1a8c9e4e126f9cd38ac3a938f973f71e0

* Updated from global requirements

Change-Id: I1d7434ca84763893c6c3a3d600d9493600d7c6e2

* Updated from global requirements

Change-Id: Ibc588572ebdd4de7dba3bf455a253ed9fdb0c89c

* Updated from global requirements

Change-Id: I6c1a8bbd9604757b265e392e47bb2ffe44052124

* Added support for inline meta variable definitions

Before, a meta variable of the form "application/|type|" would
throw an error, now you can specify |type| as a metavar.

Change-Id: I0f9ba4635cdd6e2886d7f2f5e1ce2ab5b8dc2330

* Clarified docs and error messages

Change-Id: Ic30f6ade79461770a5283b4783aaa6d547049cff

* Improve performance by multithreading test calls

This change:
1) rewrites the runner to spawn a thread pool for each template
   and assigns a worker for each test case
2) makes the output colorized by default
3) makes minor changes to the output

Change-Id: I49906f5daaa339ca9429913680203c762a0ad9fe

* Update the documentation link for doc migration

This patch is proposed according to the Direction 10 of doc
migration(https://etherpad.openstack.org/p/doc-migration-tracking).

Change-Id: I5aa1dc8952ac55de04195af51d2b0b1aeaf7a98d

* Updated from global requirements

Change-Id: Ib132ef861a517eac6609578b163c88d755534440

* Remove explicit version info from setup.cfg

Version info can be retrieved from git history.

Change-Id: I9a06017ec81aab2836a9b71febab8cecf7c1cb0c
Closes-bug: #1729547

* Fixed Broken Payload Downloading

Change-Id: I88a610bcd0855bf25c623655b4daa49276399db2
Closes-bug: #1729797

* Fixed logging issue in payload downloading

Change-Id: Ice45ca913a60ef5798fbc1d0d4e741083f647316

* Updated from global requirements

Change-Id: I360c150e7a7bb2a1f43c4fca6637835cec9810dd

* removed explicit version again

Change-Id: I7bfcde9910631678bc7d424278f6126a7cd6070b

* Update home-page url

The url of home-page is out of date. We need update it.

Change-Id: I101b7a1b7392cff5a499d77f4bf02abdb394a9a8

* Updated from global requirements

Change-Id: Iac6ba516b82cf303bca3bc204b31d5b90df54a22

* fix link

Change-Id: Ib33004eba6b41fe11a70a4e44ee10ac94e48ac56

* import zuul job settings from project-config

This is a mechanically generated patch to complete step 1 of moving
the zuul job settings out of project-config and into each project
repository.

Because there will be a separate patch on each branch, the branch
specifiers for branch-specific jobs have been removed.

Because this patch is generated by a script, there may be some
cosmetic changes to the layout of the YAML file(s) as the contents are
normalized.

See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

Change-Id: I74425c71b727d083beaa462a48bd4e442c719cd0
Story: #2002586
Task: #24760

* switch documentation job to new PTI

This is a mechanically generated patch to switch the documentation
jobs to use the new PTI versions of the jobs as part of the
python3-first goal.

See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

Change-Id: I3d533353915b828a56d219da7ee328aaa8fb1cf9
Story: #2002586
Task: #24760

* add python 3.6 unit test job

This is a mechanically generated patch to add a unit test job running
under Python 3.6 as part of the python3-first goal.

See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

Change-Id: I0fbb75bec65431d861dcf5a4dcda78bd740e1807
Story: #2002586
Task: #24760

* fix tox python3 overrides

We want to default to running all tox environments under python 3, so
set the basepython value in each environment.

We do not want to specify a minor version number, because we do not
want to have to update the file every time we upgrade python.

We do not want to set the override once in testenv, because that
breaks the more specific versions used in default environments like
py35 and py36.

Change-Id: Ia0c9884ea6c20553d2b4da8b0668b64b53ace41f
Signed-off-by: Doug Hellmann <doug@doughellmann.com>

* YAML support, basicauth extension, bugfixes

 1) Allows for YAML body in request templates. If a content-type
    is specified for a request template, Syntribos will validate the
    body against the header. This is to prevent templates from silently
    failing to parse and sending unintended data to the target.
 2) Added extension to support basicauth
 3) Lowered confidence ratings for various tests

Change-Id: I672b0e0aa3da1aa6dd7e9a8456da73f0a15759b7

* Bug fixes related to python 3 changes

Change-Id: I048bb19cf456281521883eb1500832343d72fb97

* modify keystone spelling errors

Change-Id: I3ef263c3ea78b803739f1871175d417003252430

* Correct/update links

Change-Id: I37cc088bdb579232966c4f8905ab447665d0dc54

* Bug fixes and error message updates

Fixes:
1) Crashes in runner and file_utils
2) Binary strings being read in as payloads

Updates:
1) Clarified error messages in parser
2) Confusing variable names in test cases vs issues

Adds:
1) A `syntribos root` CLI sub command to display the current syntribos root dir

Change-Id: I22edf7a1f3d39724522aee88d08b00d299b67248

* Custom root configuration option

Instead of syntribos taking in ``--custom_install_root` during init and
``--syntribos-custom_root`` during run, which sometimes led to the custom root
option not being set correctly, the redundant config options have been
deprecated.

Change-Id: I95f4d345fe8a8608ff6d1daaf06ff49912917a3f

* OpenDev Migration Patch

This commit was bulk generated and pushed by the OpenDev sysadmins
as a part of the Git hosting and code review systems migration
detailed in these mailing list posts:

http://lists.openstack.org/pipermail/openstack-discuss/2019-March/003603.html
http://lists.openstack.org/pipermail/openstack-discuss/2019-April/004920.html

Attempts have been made to correct repository namespaces and
hostnames based on simple pattern matching, but it's possible some
were updated incorrectly or missed entirely. Please reach out to us
via the contact information listed at https://opendev.org/ with any
questions you may have.

Co-authored-by: Luong Anh Tuan <tuanla@vn.fujitsu.com>
Co-authored-by: Ngo Quoc Cuong <cuongnq@vn.fujitsu.com>
Co-authored-by: OpenStack Proposal Bot <openstack-infra@lists.openstack.org>
Co-authored-by: Jenkins <jenkins@review.openstack.org>
Co-authored-by: Michael Dong <michael.dong@rackspace.com>
Co-authored-by: loooosy <syluo5695@fiberhome.com>
Co-authored-by: Jeremy Liu <liujiong@gohighsec.com>
Co-authored-by: Zuul <zuul@review.openstack.org>
Co-authored-by: melissaml <ma.lei@99cloud.net>
Co-authored-by: wangqi <wang.qi@99cloud.net>
Co-authored-by: Doug Hellmann <doug@doughellmann.com>
Co-authored-by: sunyandi <sunyd@inspur.com>
Co-authored-by: lidong <lidongbj@inspur.com>